Add tests to show 834
